requirements.txt
boto3>=1.15.11
locopy>=0.3.7
numpy>=1.19.2
pandas>=1.1.0
psutil>=5.7.0
psycopg2-binary>=2.8.0
scikit-learn>=0.23.1
sqlalchemy>=1.3.17
